pyproject.toml
[build-system]
requires = ['poetry-core']
build-backend = 'poetry.core.masonry.api'

[tool.poetry]
name = 'template_env'
version = '1.6.1'
description = 'Python dependency management and packaging made easy.'
authors = ['ds_team']

[tool.poetry.dependencies]
python = '~3.9'

[tool.poetry.group.dev.dependencies]
google-auth = "^2.25.2"
google-api-python-client = "^2.110.0"
google-auth-oauthlib = "^1.1.0"
google-cloud-bigquery-storage = "^2.23.0"
google-cloud-storage = "^2.13.0"
google-cloud-bigquery = "^3.14.0"
tqdm = "^4.66.1"
pyarrow = "^14.0.1"
openpyxl = "^3.1.2"
pandas = "^2.1.4"

[tool.poetry.group.linting.dependencies]
ruff = '^0.0.285'
isort = '^5.12.0'
black = '^23.7.0'

[tool.poetry.group.testing.dependencies]
pytest = '^7.4.0'
coverage = '^7.3.0'
mypy = '^1.5.1'
pre-commit = '^3.3.3'

[tool.ruff]
# Read more: 'https://beta.ruff.rs/docs/rules/'
ignore = ['D104', 'D100', 'D103', 'G004', 'N999', 'I001', 'T201']
select = [
    'B', # flake8-bugbear
    'C4', # flake8-comprehensions
    'ERA', # flake8-eradicate/eradicate
    'PIE', # flake8-pie
    'SIM', # flake8-simplify
    'TCH', # flake8-type-checking
    'TID', # flake8-tidy-imports
    'PTH', # flake8-use-pathlib
    'ARG', # flake8-unused-arguments
    'Q', # flake8-quotes
    'PT', # flake8-pytest-style
    'T20', # flake8-print
    'G', # flake8-logging-format
    'ICN', # flake8-import-conventions
    'COM', # flake8-commas
    'A', # flake8-builtins
    'D', # pydocstyle
    'I', # isort
    'N', # pep8-naming
    'RUF', # ruff checks
    'FLY', # flynt
    'PD', # pandas-vet
]

# Exclude a variety of commonly ignored directories.
exclude = [
    '.git',
    '.gitlab',
    '.mypy_cache',
    '.pytype',
    '.ruff_cache',
    '.venv',
    '__pypackages__',
    'venv',
]

line-length = 88 # Same as Black

[tool.ruff.flake8-quotes]
docstring-quotes = 'double'
inline-quotes = 'single'
multiline-quotes = 'single'

[tool.black]
profile = "black"
line-length = 88
skip-string-normalization = true

[tool.isort] # Making isort compatible with black
multi_line_output = 3
line_length = 88
include_trailing_comma = true

[tool.ruff.pydocstyle]
convention = 'google'

[tool.coverage.report]
skip_covered = true
fail_under = 80
omit = ['utils/*', 'config/*']

[tool.mypy]
ignore_missing_imports = true
exclude = ['tests', 'config']